Modify db models (#151)
